Updated structure and added support for nested components. Action to rename components
