+++ client/packages/lowcoder/src/comps/comps/preLoadComp/actionConfigs.ts
@@ -1,573 +1,27 @@
-import { message } from "antd";
-import { genRandomKey } from "comps/utils/idGenerator";
-import { parseCompType } from "comps/utils/remote";
-import { defaultLayout, GridItemDataType } from "comps/comps/gridItemComp";
-import { addMapChildAction } from "comps/generators/sameTypeMap";
-import { uiCompRegistry, UICompType } from "comps/uiCompRegistry";
-import { ActionConfig, ActionCategory, ActionExecuteParams } from "./types";
-import { 
-    multiChangeAction, 
-    wrapActionExtraInfo, 
-    changeValueAction, 
-    wrapChildAction, 
-    deleteCompAction 
-} from "lowcoder-core";
-
-// Component Management Actions
-const addComponentAction: ActionConfig = {
-  key: 'add-components',
-  label: 'Place a component',
-  category: 'component-management',
-  requiresComponentSelection: true,
-  requiresInput: false,
-  execute: async (params: ActionExecuteParams) => {
-    const { selectedComponent, editorState } = params;
-    
-    if (!selectedComponent || !editorState) {
-      message.error('Component and editor state are required');
-      return;
-    }
-
-    try {
-      // Get the UI component container
-      const uiComp = editorState.getUIComp();
-      const container = uiComp.getComp();
-      
-      if (!container) {
-        message.error('No container available to add component');
-        return;
-      }
-
-      // Get the simple container (the actual grid container)
-      const simpleContainer = container.realSimpleContainer();
-      if (!simpleContainer) {
-        message.error('No grid container available');
-        return;
-      }
-
-      // Generate component name and key
-      const nameGenerator = editorState.getNameGenerator();
-      const compInfo = parseCompType(selectedComponent);
-      const compName = nameGenerator.genItemName(compInfo.compName);
-      const key = genRandomKey();
-
-      // Get component manifest and default data
-      const manifest = uiCompRegistry[selectedComponent];
-      let defaultDataFn = undefined;
-
-      if (manifest?.lazyLoad) {
-        const { defaultDataFnName, defaultDataFnPath } = manifest;
-        if (defaultDataFnName && defaultDataFnPath) {
-          const module = await import(`../../${defaultDataFnPath}.tsx`);
-          defaultDataFn = module[defaultDataFnName];
-        }
-      } else if (!compInfo.isRemote) {
-        defaultDataFn = manifest?.defaultDataFn;
-      }
-
-      // Create component data
-      const widgetValue: GridItemDataType = {
-        compType: selectedComponent,
-        name: compName,
-        comp: defaultDataFn ? defaultDataFn(compName, nameGenerator, editorState) : undefined,
-      };
-
-      // Get current layout and calculate position
-      const currentLayout = simpleContainer.children.layout.getView();
-      const layoutInfo = manifest?.layoutInfo || defaultLayout(selectedComponent as UICompType);
-      
-      let itemPos = 0;
-      if (Object.keys(currentLayout).length > 0) {
-        itemPos = Math.min(...Object.values(currentLayout).map((l: any) => l.pos || 0)) - 1;
-      }
-
-      const layoutItem = {
-        i: key,
-        x: 0,
-        y: 0,
-        w: layoutInfo.w || 6,
-        h: layoutInfo.h || 5,
-        pos: itemPos,
-        isDragging: false,
-      };
-
-      // Dispatch the action to add the component
-      simpleContainer.dispatch(
-        wrapActionExtraInfo(
-          multiChangeAction({
-            layout: changeValueAction(
-              {
-                ...currentLayout,
-                [key]: layoutItem,
-              },
-              true
-            ),
-            items: addMapChildAction(key, widgetValue),
-          }),
-          { compInfos: [{ compName: compName, compType: selectedComponent, type: "add" }] }
-        )
-      );
-
-      editorState.setSelectedCompNames(new Set([compName]), "addComp");
-      
-      message.success(`Component "${manifest?.name || selectedComponent}" added successfully!`);
-    } catch (error) {
-      console.error('Error adding component:', error);
-      message.error('Failed to add component. Please try again.');
-    }
-  }
-};
-
-const moveComponentAction: ActionConfig = {
-  key: 'move-components',
-  label: 'Move a component',
-  category: 'component-management',
-  requiresEditorComponentSelection: true,
-  requiresInput: true,
-  inputPlaceholder: 'Enter move parameters (e.g., x:100, y:200)',
-  inputType: 'text',
-  validation: (value: string) => {
-    if (!value.trim()) return 'Move parameters are required';
-    
-    // Parse the input to validate format
-    const params = value.toLowerCase().split(',').map(p => p.trim());
-    for (const param of params) {
-      if (!param.includes(':')) {
-        return 'Invalid format. Use "x:value, y:value"';
-      }
-      const [key, val] = param.split(':').map(s => s.trim());
-      if (!['x', 'y'].includes(key)) {
-        return 'Only x and y parameters are supported';
-      }
-      const num = parseInt(val);
-      if (isNaN(num) || num < 0) {
-        return `${key} must be a positive number`;
-      }
-    }
-    return null;
-  },
-  execute: async (params: ActionExecuteParams) => {
-    const { selectedEditorComponent, actionValue, editorState } = params;
-
-    if (!selectedEditorComponent || !editorState) {
-      message.error('Component and editor state are required');
-      return;
-    }
-
-    try {
-      // Parse move parameters
-      const moveParams: { x?: number; y?: number } = {};
-      const params = actionValue.toLowerCase().split(',').map(p => p.trim());
-      
-      for (const param of params) {
-        const [key, val] = param.split(':').map(s => s.trim());
-        if (['x', 'y'].includes(key)) {
-          moveParams[key as 'x' | 'y'] = parseInt(val);
-        }
-      }
-
-      if (!moveParams.x && !moveParams.y) {
-        message.error('No valid move parameters provided');
-        return;
-      }
-
-      // Get the UI component container
-      const uiComp = editorState.getUIComp();
-      const container = uiComp.getComp();
-      
-      if (!container) {
-        message.error('No container available');
-        return;
-      }
-
-      // Get the simple container (the actual grid container)
-      const simpleContainer = container.realSimpleContainer();
-      if (!simpleContainer) {
-        message.error('No grid container available');
-        return;
-      }
-
-      // Get current layout and items
-      const currentLayout = simpleContainer.children.layout.getView();
-      const items = simpleContainer.children.items.children;
-
-      // Find the component by name and get its key
-      let componentKey: string | null = null;
-      for (const [key, item] of Object.entries(items)) {
-        if ((item as any).children.name.getView() === selectedEditorComponent) {
-          componentKey = key;
-          break;
-        }
-      }
-
-      if (!componentKey || !currentLayout[componentKey]) {
-        message.error(`Component "${selectedEditorComponent}" not found in layout`);
-        return;
-      }
-
-      const currentLayoutItem = currentLayout[componentKey];
-
-      const newLayoutItem = {
-        ...currentLayoutItem,
-        x: moveParams.x !== undefined ? moveParams.x : currentLayoutItem.x,
-        y: moveParams.y !== undefined ? moveParams.y : currentLayoutItem.y,
-      };
-
-      const newLayout = {
-        ...currentLayout,
-        [componentKey]: newLayoutItem,
-      };
-
-      // Dispatch the action to update the layout
-      simpleContainer.dispatch(
-        wrapActionExtraInfo(
-          multiChangeAction({
-            layout: changeValueAction(newLayout, true),
-          }),
-          { 
-            compInfos: [{ 
-              compName: selectedEditorComponent, 
-              compType: (items[componentKey] as any).children.compType.getView(), 
-              type: "layout" 
-            }] 
-          }
-        )
-      );
-
-      editorState.setSelectedCompNames(new Set([selectedEditorComponent]), "moveComp");
-      
-      const moveDescription = [];
-      if (moveParams.x !== undefined) moveDescription.push(`x: ${moveParams.x}`);
-      if (moveParams.y !== undefined) moveDescription.push(`y: ${moveParams.y}`);
-      
-      message.success(`Component "${selectedEditorComponent}" moved to ${moveDescription.join(', ')}`);
-    } catch (error) {
-      console.error('Error moving component:', error);
-      message.error('Failed to move component. Please try again.');
-    }
-  }
-};
-
-const deleteComponentAction: ActionConfig = {
-  key: 'delete-components',
-  label: 'Delete a component',
-  category: 'component-management',
-  requiresEditorComponentSelection: true,
-  requiresInput: false,
-  execute: async (params: ActionExecuteParams) => {
-    const { selectedEditorComponent, editorState } = params;
-    
-    if (!selectedEditorComponent || !editorState) {
-      message.error('Component and editor state are required');
-      return;
-    }
-
-    try {
-      // Get the UI component container
-      const uiComp = editorState.getUIComp();
-      const container = uiComp.getComp();
-      
-      if (!container) {
-        message.error('No container available');
-        return;
-      }
-
-      // Get the simple container (the actual grid container)
-      const simpleContainer = container.realSimpleContainer();
-      if (!simpleContainer) {
-        message.error('No grid container available');
-        return;
-      }
-
-      // Get current layout and items
-      const currentLayout = simpleContainer.children.layout.getView();
-      const items = simpleContainer.children.items.children;
-
-      // Find the component by name and get its key
-      let componentKey: string | null = null;
-      let componentType: string | null = null;
-      for (const [key, item] of Object.entries(items)) {
-        if ((item as any).children.name.getView() === selectedEditorComponent) {
-          componentKey = key;
-          componentType = (item as any).children.compType.getView();
-          break;
-        }
-      }
-
-      if (!componentKey || !currentLayout[componentKey]) {
-        message.error(`Component "${selectedEditorComponent}" not found in layout`);
-        return;
-      }
-
-      // Create new layout without the component
-      const newLayout = { ...currentLayout };
-      delete newLayout[componentKey];
-
-      // Dispatch the action to remove the component
-      simpleContainer.dispatch(
-        wrapActionExtraInfo(
-          multiChangeAction({
-            layout: changeValueAction(newLayout, true),
-            items: wrapChildAction(componentKey, deleteCompAction()),
-          }),
-          { 
-            compInfos: [{ 
-              compName: selectedEditorComponent, 
-              compType: componentType || 'unknown', 
-              type: "delete" 
-            }] 
-          }
-        )
-      );
-
-      editorState.setSelectedCompNames(new Set(), "deleteComp");
-      
-      message.success(`Component "${selectedEditorComponent}" deleted successfully`);
-    } catch (error) {
-      console.error('Error deleting component:', error);
-      message.error('Failed to delete component. Please try again.');
-    }
-  }
-};
-
-const resizeComponentAction: ActionConfig = {
-  key: 'resize-components',
-  label: 'Resize a component',
-  category: 'component-management',
-  requiresEditorComponentSelection: true,
-  requiresInput: true,
-  inputPlaceholder: 'Enter resize parameters (e.g., w:8, h:6)',
-  inputType: 'text',
-  validation: (value: string) => {
-    if (!value.trim()) return 'Resize parameters are required';
-    
-    // Parse the input to validate format
-    const params = value.toLowerCase().split(',').map(p => p.trim());
-    for (const param of params) {
-      if (!param.includes(':')) {
-        return 'Invalid format. Use "w:value, h:value"';
-      }
-      const [key, val] = param.split(':').map(s => s.trim());
-      if (!['w', 'h'].includes(key)) {
-        return 'Only w (width) and h (height) parameters are supported';
-      }
-      const num = parseInt(val);
-      if (isNaN(num) || num < 1) {
-        return `${key} must be a positive number greater than 0`;
-      }
-    }
-    return null;
-  },
-  execute: async (params: ActionExecuteParams) => {
-    const { selectedEditorComponent, actionValue, editorState } = params;
-
-    if (!selectedEditorComponent || !editorState) {
-      message.error('Component and editor state are required');
-      return;
-    }
-
-    try {
-      // Parse resize parameters
-      const resizeParams: { w?: number; h?: number } = {};
-      const params = actionValue.toLowerCase().split(',').map(p => p.trim());
-      
-      for (const param of params) {
-        const [key, val] = param.split(':').map(s => s.trim());
-        if (['w', 'h'].includes(key)) {
-          resizeParams[key as 'w' | 'h'] = parseInt(val);
-        }
-      }
-
-      if (!resizeParams.w && !resizeParams.h) {
-        message.error('No valid resize parameters provided');
-        return;
-      }
-
-      // Get the UI component container
-      const uiComp = editorState.getUIComp();
-      const container = uiComp.getComp();
-      
-      if (!container) {
-        message.error('No container available');
-        return;
-      }
-
-      // Get the simple container (the actual grid container)
-      const simpleContainer = container.realSimpleContainer();
-      if (!simpleContainer) {
-        message.error('No grid container available');
-        return;
-      }
-
-      // Get current layout and items
-      const currentLayout = simpleContainer.children.layout.getView();
-      const items = simpleContainer.children.items.children;
-
-      // Find the component by name and get its key
-      let componentKey: string | null = null;
-      for (const [key, item] of Object.entries(items)) {
-        if ((item as any).children.name.getView() === selectedEditorComponent) {
-          componentKey = key;
-          break;
-        }
-      }
-
-      if (!componentKey || !currentLayout[componentKey]) {
-        message.error(`Component "${selectedEditorComponent}" not found in layout`);
-        return;
-      }
-
-      const currentLayoutItem = currentLayout[componentKey];
-      
-      const newLayoutItem = {
-        ...currentLayoutItem,
-        w: resizeParams.w !== undefined ? resizeParams.w : currentLayoutItem.w,
-        h: resizeParams.h !== undefined ? resizeParams.h : currentLayoutItem.h,
-      };
-
-      const newLayout = {
-        ...currentLayout,
-        [componentKey]: newLayoutItem,
-      };
-
-      // Dispatch the action to update the layout
-      simpleContainer.dispatch(
-        wrapActionExtraInfo(
-          multiChangeAction({
-            layout: changeValueAction(newLayout, true),
-          }),
-          { 
-            compInfos: [{ 
-              compName: selectedEditorComponent, 
-              compType: (items[componentKey] as any).children.compType.getView(), 
-              type: "layout" 
-            }] 
-          }
-        )
-      );
-
-      editorState.setSelectedCompNames(new Set([selectedEditorComponent]), "resizeComp");
-      
-      const resizeDescription = [];
-      if (resizeParams.w !== undefined) resizeDescription.push(`width: ${resizeParams.w}`);
-      if (resizeParams.h !== undefined) resizeDescription.push(`height: ${resizeParams.h}`);
-      
-      message.success(`Component "${selectedEditorComponent}" resized to ${resizeDescription.join(', ')}`);
-    } catch (error) {
-      console.error('Error resizing component:', error);
-      message.error('Failed to resize component. Please try again.');
-    }
-  }
-};
-
-// Component Configuration Actions
-const configureComponentAction: ActionConfig = {
-  key: 'configure-components',
-  label: 'Configure a component',
-  category: 'component-configuration',
-  requiresEditorComponentSelection: true,
-  requiresInput: true,
-  inputPlaceholder: 'Enter configuration (JSON format)',
-  inputType: 'json',
-  validation: (value: string) => {
-    if (!value.trim()) return 'Configuration is required';
-    try {
-      JSON.parse(value);
-      return null;
-    } catch {
-      return 'Invalid JSON format';
-    }
-  },
-  execute: async (params: ActionExecuteParams) => {
-    const { selectedEditorComponent, actionValue } = params;
-    
-    try {
-      const config = JSON.parse(actionValue);
-      console.log('Configuring component:', selectedEditorComponent, 'with config:', config);
-      message.info(`Configure action for component "${selectedEditorComponent}"`);
-      
-      // TODO: Implement actual configuration logic
-    } catch (error) {
-      message.error('Invalid configuration format');
-    }
-  }
-};
-
-// Layout Actions
-const changeLayoutAction: ActionConfig = {
-  key: 'change-layout',
-  label: 'Change layout',
-  category: 'layout',
-  requiresInput: true,
-  inputPlaceholder: 'Enter layout type (grid, flex, absolute)',
-  inputType: 'text',
-  execute: async (params: ActionExecuteParams) => {
-    const { actionValue } = params;
-    
-    console.log('Changing layout to:', actionValue);
-    message.info(`Layout changed to: ${actionValue}`);
-    
-    // TODO: Implement actual layout change logic
-  }
-};
-
-// Event Actions
-const addEventHandlerAction: ActionConfig = {
-  key: 'add-event-handler',
-  label: 'Add event handler',
-  category: 'events',
-  requiresEditorComponentSelection: true,
-  requiresInput: true,
-  inputPlaceholder: 'Enter event handler code (JavaScript)',
-  inputType: 'textarea',
-  execute: async (params: ActionExecuteParams) => {
-    const { selectedEditorComponent, actionValue } = params;
-    
-    console.log('Adding event handler to component:', selectedEditorComponent, 'with code:', actionValue);
-    message.info(`Event handler added to component "${selectedEditorComponent}"`);
-    
-    // TODO: Implement actual event handler logic
-  }
-};
-
-// Style Actions
-const applyStyleAction: ActionConfig = {
-  key: 'apply-style',
-  label: 'Apply style to component',
-  category: 'styling',
-  requiresEditorComponentSelection: true,
-  requiresInput: true,
-  inputPlaceholder: 'Enter CSS styles (JSON format)',
-  inputType: 'json',
-  validation: (value: string) => {
-    if (!value.trim()) return 'Styles are required';
-    try {
-      JSON.parse(value);
-      return null;
-    } catch {
-      return 'Invalid JSON format';
-    }
-  },
-  execute: async (params: ActionExecuteParams) => {
-    const { selectedEditorComponent, actionValue } = params;
-    
-    try {
-      const styles = JSON.parse(actionValue);
-      console.log('Applying styles to component:', selectedEditorComponent, 'with styles:', styles);
-      message.info(`Styles applied to component "${selectedEditorComponent}"`);
-      
-      // TODO: Implement actual style application logic
-    } catch (error) {
-      message.error('Invalid style format');
-    }
-  }
-};
+import { ActionCategory } from "./types";
+import {
+  addComponentAction,
+  moveComponentAction,
+  renameComponentAction,
+  deleteComponentAction,
+  resizeComponentAction,
+  configureComponentAction,
+  changeLayoutAction,
+  addEventHandlerAction,
+  applyStyleAction
+} from "./actions";
 
 export const actionCategories: ActionCategory[] = [
   {
     key: 'component-management',
     label: 'Component Management',
-    actions: [addComponentAction, moveComponentAction, deleteComponentAction, resizeComponentAction]
+    actions: [
+      addComponentAction, 
+      moveComponentAction, 
+      deleteComponentAction, 
+      resizeComponentAction,
+      renameComponentAction
+    ]
   },
   {
     key: 'component-configuration',
@@ -591,7 +45,7 @@ export const actionCategories: ActionCategory[] = [
   }
 ];
 
-export const actionRegistry = new Map<string, ActionConfig>();
+export const actionRegistry = new Map<string, any>();
 actionCategories.forEach(category => {
   category.actions.forEach(action => {
     actionRegistry.set(action.key, action);
@@ -617,4 +71,4 @@ export const getAllActionItems = () => {
         }))
       }];
     });
-  };
+  };
