+++ client/packages/lowcoder/src/comps/comps/preLoadComp/actions/index.ts
@@ -1,0 +1,14 @@
+// Component Management Actions
+export * from './componentManagement';
+
+// Component Configuration Actions
+export { configureComponentAction } from './componentConfiguration';
+
+// Layout Actions
+export { changeLayoutAction } from './componentLayout';
+
+// Event Actions
+export { addEventHandlerAction } from './componentEvents';
+
+// Styling Actions
+export { applyStyleAction } from './componentStyling';
